app/requirements.txt
Telethon==1.22.0
selenium==2.53.6
pysocks
beautifulsoup4==4.9.3
lxml
wget